OvmfPkg: align sections to 4K page boundaries on XCODE toolchains
In addition to the linker flags, XCODE toolchains use the mtoc utility
to convert Mach-O files to PECOFF files. The alignment needs to be
specified to both. Additionally, the linker will warn that the starting
address is not on a page boundary, so it is also set explicitly to one.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Will Ross <paxswill@paxswill.com>